requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
streamlit
requests
